+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 List all changes after the last release here (newer on top). Each change on a separate bullet point line
 
+### Added
+
+- Admin: add a method in AdminModule that allows returning help texts for permissions
+
 ### Fixed
 
 - Importer: use more bites in order to detect csv dialect in importer